Replace `initializer...def load_console` business with a simple `console do` block
`instance_eval` is a bit too hacky to my taste. There's more natural way
to handle this, namely the `::Rails::Railtie.console` method. It is not
quite well-documented, but it is public and does the job. It works
pretty much the same way as `.generators` and `.rake_tasks` methods
(those are documented).

* https://github.com/rails/rails/blob/v4.1.2.rc1/railties/lib/rails/railtie.rb#L141
* http://api.rubyonrails.org/classes/Rails/Railtie.html#method-i-generators